coloquei alguns comentarios de debug aparentemente só funciona nas portas 16 e 5

+++ main/main.cpp
@@ -39,7 +39,7 @@ void ajustaSetPoint()
         printf("Digite o novo set point: ");
         char newSetPointStr[3];
         serial.readString((uint8_t *)newSetPointStr, 3);
-        int newSetPoint = atof(newSetPointStr);
+        int newSetPoint = atoi(newSetPointStr);
         printf("Novo set point: %d\n", newSetPoint);
         sensor.setupBaseHeight(newSetPoint);
     }
